+++ packages/import/i18n/en.js
@@ -3,6 +3,7 @@ i18n.map('en', {
 	label_import: "Import Data",
 	label_importfile: "Upload a GeoJSON File",
 	label_importname: "Import Name",
+	label_importdata: "Import Data",
 	label_importpreview: "Count of importable places",
 	label_imported: "Imported places",
 	label_imports: "Imports",