+++ packages/import/i18n/it.js
@@ -2,7 +2,8 @@
 i18n.map('it', {
 	label_import: "Importa Dati",
 	label_importfile: "Carica un file GeoJSON",
-	label_importname: "Nome Importazione",	
+	label_importname: "Nome Importazione",
+	label_importdata: "Dati Importazione",	
 	label_imported: "Luoghi Importati",
 	label_imports: "Importazioni",
 	error_import_formatNotValid: "Formato non Valido"