+++ packages/import/modules/Util_sanitize.js
@@ -17,5 +17,20 @@ _.extend(K.Util.sanitize, {
 			.replace(/[\-\.]/g,'');
 
 		return name;
+	},
+	/**
+	 * reduce mutiplegeometry into single one if it contains only one geometry
+	 * @param  {[type]} geom [description]
+	 * @return {[type]}      [description]
+	 */
+	importGeometry: function(geom) {
+		if(_.str.startsWith(geom.type,'Multi') && geom.coordinates.length===1) {
+			return {
+				type: geom.type.replace('Multi',''),
+				coordinates: geom.coordinates[0] 
+			}
+		}
+		else
+			return geom;
 	}
 });